Add files via upload
Python script created to check if a padding oracle attack can be performed on HTB's machine Overflow. It does register a bunch of accounts (50), and then prints usernames and cookies length in order to see if there is any relation between them. There is also another function which tries to log in, clear the cookie and set a new one, and finally perform a get request to see what happens and if any padding error message is leaked (which would lead to a potential padding oracle attack).