Add files via upload
Given a custom alphabet and a flag I needed to decrypt it. The quote on the top is a reference to Julius Caesar, and after doing some research, I found that there is a pretty simple encryption type with its name. However I didn't had the shift number, so after some guesses, I just tried to brute force it through all the 65 possibles flags. The pdb library was pretty useful, as I needed to debug my code at many points.